refactor(paywalls): split web_view schema and view tests by target
WebViewComponentTests mixed RevenueCat schema (Codable) tests with
RevenueCatUI view-model/coordinator tests in a single file. Split them to
match the Video component convention:

- Tests/UnitTests/Paywalls/Components/WebViewComponentTests.swift: schema
  decoding tests (RevenueCat only), wired into the xcodeproj UnitTests
  target so they run under `fastlane scan` again.
- Tests/RevenueCatUITests/PaywallsV2/WebViewComponentViewTests.swift:
  view-model/sizing/coordinator tests (RevenueCatUI), SPM only.
